# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Sprache Account Binding API
  version: 0.1.0
  description: Stores only an anonymized Google account to Drive folder mapping.
servers:
  - url: http://127.0.0.1:3000
paths:
  /health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: Service is running.
          content:
            application/json:
              schema:
                type: object
                required: [status, service]
                properties:
                  status:
                    const: ok
                  service:
                    const: sprache-api
  /v1/auth/verify:
    post:
      operationId: verifyAuthentication
      security:
        - googleIdToken: []
      responses:
        "200":
          description: Token is valid.
          content:
            application/json:
              schema:
                type: object
                required: [authenticated, driveConnected]
                properties:
                  authenticated:
                    const: true
                  driveConnected:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/me/drive-root:
    get:
      operationId: getDriveRoot
      security:
        - googleIdToken: []
      responses:
        "200":
          description: Current account binding, or null.
          content:
            application/json:
              schema:
                type: object
                required: [binding]
                properties:
                  binding:
                    oneOf:
                      - $ref: "#/components/schemas/DriveBinding"
                      - type: "null"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: putDriveRoot
      security:
        - googleIdToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [folderId]
              properties:
                folderId:
                  type: string
                  pattern: "^[A-Za-z0-9_-]{10,256}$"
                folderName:
                  type: [string, "null"]
                  maxLength: 256
                schemaVersion:
                  type: integer
                  minimum: 1
                  default: 1
      responses:
        "200":
          description: Binding stored.
          content:
            application/json:
              schema:
                type: object
                required: [binding]
                properties:
                  binding:
                    $ref: "#/components/schemas/DriveBinding"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteDriveRoot
      security:
        - googleIdToken: []
      responses:
        "204":
          description: Binding removed. Drive files are not deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    googleIdToken:
      type: http
      scheme: bearer
      bearerFormat: Google ID Token
  schemas:
    DriveBinding:
      type: object
      required:
        [folderId, folderName, schemaVersion, createdAt, updatedAt]
      properties:
        folderId:
          type: string
        folderName:
          type: [string, "null"]
        schemaVersion:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
  responses:
    Unauthorized:
      description: Missing, invalid, or expired Google ID Token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InvalidRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"